Target [Helldar\LaravelLangPublisher\Contracts\Pathable] is not instantiable.

+++ src/Services/Processors/DeleteJson.php
@@ -13,6 +13,10 @@ final class DeleteJson extends BaseProcessor
 
     protected function delete(): bool
     {
+        if (! IlluminateFile::exists($this->targetPath())) {
+            return true;
+        }
+
         return ! $this->isProtected()
             ? IlluminateFile::delete($this->targetPath())
             : false;
